admin-ui: split ehr_detail.rs into a per-tab directory module
The 1590-line pages/ehr_detail.rs is replaced by a pages/ehr_detail/
directory: mod.rs keeps the EhrDetailPage shell, the shared tab_bar, and
the commit_version_uid helper shared by the directory and composition
commit paths; one sibling file per tab holds that tab's #[server] fns,
view sections, and inline tests — status.rs, directory.rs (incl. the
folder-template store types), compositions.rs, contributions.rs.

Pure relocation: no view, styling, id, or logic change. Cross-module
callers import each item from its new defining module (zero re-exports):
folder_templates.rs now points at ehr_detail::directory for
DIRECTORY_ARCHETYPE/FOLDER_NODE_ID/FolderTemplate/folder_json. Items used
only within their new file are narrowed to private (empty_root_folder,
parse_contributions); commit_version_uid stays private in the parent and
is reached by the descendant tab modules.
