--- app/ehrbase-admin-ui/src/pages/ehr_detail.rs
@@ -1,1590 +1,0 @@
-//! The `/ehrs/{ehr_id}` screen — EHR detail: status / directory / compositions /
-//! contributions tabs.
-//!
-//! Four URL-driven tabs (`?tab=`, rules §9) over one EHR. Each tab's data is a
-//! `#[server]` fn co-located here; the resources are created once and their
-//! sources are gated on the active tab (a `Memo` over the query map), so only
-//! the visible tab fetches (rules §6 — never fetch-in-effect). The tab bodies
-//! are always mounted and toggled with `class:hidden`, keeping the server and
-//! client view structure identical (rules §8 — no `cfg!`-branched structure).
-//!
-//! No openEHR spec governs an admin UI — our own design / product extension.
-//! The wire it reads IS spec-bound (ITS-REST EHR + Query APIs). User input
-//! NEVER concatenates into AQL — the fixed query is a validated const and the
-//! `ehr_id` travels as an AQL `query_parameters` binding; path segments are
-//! percent-encoded server-side.
-//!
-//! Every co-located `#[server]` fn guards with
-//! [`require_session`](crate::session::require_session) first (rules §0), and
-//! the CDR credential never reaches client-visible state.
-
-use leptos::prelude::*;
-use leptos::{component, server};
-use leptos_meta::Title;
-use leptos_router::components::A;
-use serde::{Deserialize, Serialize};
-use serde_json::{Value, json};
-
-use crate::components::data_table::{CELL, CELL_MONO, ROW, table_shell};
-use crate::components::empty_state::EmptyState;
-use crate::components::field::{BTN_PRIMARY, BTN_SECONDARY, INPUT, LABEL, SELECT, TEXTAREA};
-use crate::components::format_view::DocumentPane;
-use crate::components::page_header::{Crumb, PageHeader};
-use crate::components::surface::{CARD_PAD, CARD_TITLE, WELL};
-use crate::components::toast::toast_success;
-use crate::error::AdminUiError;
-use crate::format::ReprFormat;
-use crate::pages::ehrs::{ResultPage, cell_text, paging_controls, table_skeleton};
-// Server-side helpers, compiled only where the #[server] bodies exist.
-#[cfg(feature = "ssr")]
-use crate::pages::ehrs::{aql_request_body, parse_result_set};
-
-#[cfg(feature = "ssr")]
-/// The fixed AQL that lists an EHR's compositions newest-first. The `ehr_id`
-/// is bound as an AQL parameter (`$ehr_id`), never string-interpolated.
-/// Validated by [`tests::fixed_aql_parses`].
-const LIST_COMPOSITIONS_AQL: &str = "SELECT c/uid/value, c/name/value, \
-c/archetype_details/template_id/value, c/context/start_time/value \
-FROM EHR e CONTAINS COMPOSITION c WHERE e/ehr_id/value = $ehr_id \
-ORDER BY c/context/start_time/value DESC";
-
-/// The EHR's `EHR_STATUS` resource, raw canonical JSON.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session; CDR transport
-/// errors pass through; a non-2xx CDR answer normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn fetch_ehr_status(ehr_id: String) -> Result<String, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let url = state
-        .cdr
-        .rest_v1(&format!("ehr/{}/ehr_status", urlencoding::encode(&ehr_id)));
-    let response = state
-        .cdr
-        .get(&session.credential, &url, "application/json")
-        .await?;
-    Ok(crate::cdr::CdrClient::expect_success(response)?.body)
-}
-
-/// The EHR's directory as its canonical FOLDER JSON body plus the current
-/// version uid (the FOLDER's `uid.value`, an `OBJECT_VERSION_ID`). The uid is
-/// what the update path sends in `If-Match`: [`CdrResponse`] carries no header
-/// map, so it is read from the returned FOLDER body — the FOLDER is a
-/// `VERSIONABLE` and always carries `uid` (ITS-REST
-/// `specifications/schemas/ehr/Folder.yaml`; RM common
-/// `master05-directory_package`).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct DirectoryState {
-    /// The canonical FOLDER JSON body.
-    pub body: String,
-    /// The FOLDER's `uid.value` (`OBJECT_VERSION_ID`), or empty when absent.
-    pub version_uid: String,
-}
-
-/// The EHR's directory (root `FOLDER`) as a [`DirectoryState`], or `None` when
-/// the CDR has no directory for this EHR (a `404` is a first-class empty
-/// state, not an error).
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session; CDR transport
-/// errors pass through; a non-2xx, non-404 CDR answer normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn fetch_directory(ehr_id: String) -> Result<Option<DirectoryState>, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let url = state
-        .cdr
-        .rest_v1(&format!("ehr/{}/directory", urlencoding::encode(&ehr_id)));
-    let response = state
-        .cdr
-        .get(&session.credential, &url, "application/json")
-        .await?;
-    if response.status == 404 {
-        return Ok(None);
-    }
-    let body = crate::cdr::CdrClient::expect_success(response)?.body;
-    let version_uid = commit_version_uid(&body);
-    Ok(Some(DirectoryState { body, version_uid }))
-}
-
-/// Create the EHR's directory (`POST /ehr/{ehr_id}/directory`) from a canonical
-/// JSON FOLDER `body`. `Content-Type: application/json`,
-/// `Prefer: return=representation`, `Accept: application/json`; the new
-/// version uid (`uid.value` of the returned FOLDER) is returned (empty when the
-/// CDR returned no representation). Directory operations per ITS-REST
-/// `specifications/operations/directory_create.yaml`.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session;
-/// [`AdminUiError::Invalid`] on an empty body; CDR transport errors pass
-/// through; a non-2xx CDR answer (its validation diagnostics, which the UI
-/// renders verbatim, included) normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn create_directory(ehr_id: String, body: String) -> Result<String, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    if body.trim().is_empty() {
-        return Err(AdminUiError::Invalid(
-            "the directory body is empty".to_owned(),
-        ));
-    }
-    let url = state
-        .cdr
-        .rest_v1(&format!("ehr/{}/directory", urlencoding::encode(&ehr_id)));
-    let response = state
-        .cdr
-        .post(
-            &session.credential,
-            &url,
-            "application/json",
-            "application/json",
-            &[("Prefer", "return=representation")],
-            body,
-        )
-        .await?;
-    let response = crate::cdr::CdrClient::expect_success(response)?;
-    Ok(commit_version_uid(&response.body))
-}
-
-/// Update the EHR's directory (`PUT /ehr/{ehr_id}/directory`) with a new
-/// canonical JSON FOLDER `body`. The current `version_uid` (the
-/// `preceding_version_uid`) is sent quoted in `If-Match`, per ITS-REST
-/// `specifications/operations/directory_update.yaml`;
-/// `Prefer: return=representation` yields the new version uid. A stale uid is
-/// answered `412` by the CDR and surfaces as the normalized error.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session;
-/// [`AdminUiError::Invalid`] on an empty body or missing current version uid;
-/// CDR transport errors pass through; a non-2xx CDR answer (its validation
-/// diagnostics, which the UI renders verbatim, included) normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn update_directory(
-    ehr_id: String,
-    current_version_uid: String,
-    body: String,
-) -> Result<String, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    if body.trim().is_empty() {
-        return Err(AdminUiError::Invalid(
-            "the directory body is empty".to_owned(),
-        ));
-    }
-    let current = current_version_uid.trim();
-    if current.is_empty() {
-        return Err(AdminUiError::Invalid(
-            "the current directory version uid is required to update".to_owned(),
-        ));
-    }
-    let if_match = format!("\"{current}\"");
-    let url = state
-        .cdr
-        .rest_v1(&format!("ehr/{}/directory", urlencoding::encode(&ehr_id)));
-    let response = state
-        .cdr
-        .put(
-            &session.credential,
-            &url,
-            "application/json",
-            "application/json",
-            &[
-                ("If-Match", if_match.as_str()),
-                ("Prefer", "return=representation"),
-            ],
-            body,
-        )
-        .await?;
-    let response = crate::cdr::CdrClient::expect_success(response)?;
-    Ok(commit_version_uid(&response.body))
-}
-
-/// List the console-local folder templates (built-in defaults when the store
-/// file is absent — see [`crate::folder_templates`]).
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session;
-/// [`AdminUiError::Internal`] on an unreadable templates file.
-#[server]
-pub async fn list_folder_templates() -> Result<Vec<FolderTemplate>, AdminUiError> {
-    crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let path = crate::folder_templates::templates_path(&state.config.groups_file());
-    // File I/O off the async runtime (reliability rule: no sync I/O on it).
-    tokio::task::spawn_blocking(move || crate::folder_templates::read_templates(&path))
-        .await
-        .map_err(|e| AdminUiError::Internal(format!("folder-templates task: {e}")))?
-}
-
-/// Save a console-local folder template `name` from a canonical FOLDER JSON
-/// `folder_body` (e.g. the EHR's current directory). Upserts by name.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session;
-/// [`AdminUiError::Invalid`] on an empty name or non-JSON body;
-/// [`AdminUiError::Internal`] on an unwritable templates file.
-#[server]
-pub async fn save_folder_template(name: String, folder_body: String) -> Result<(), AdminUiError> {
-    crate::session::require_session().await?;
-    let name = name.trim().to_owned();
-    if name.is_empty() {
-        return Err(AdminUiError::Invalid(
-            "the folder template needs a name".to_owned(),
-        ));
-    }
-    let folder: Value = serde_json::from_str(&folder_body).map_err(|e| {
-        AdminUiError::Invalid(format!("the folder template is not valid JSON: {e}"))
-    })?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let path = crate::folder_templates::templates_path(&state.config.groups_file());
-    tokio::task::spawn_blocking(move || {
-        crate::folder_templates::write_template(&path, &name, folder)
-    })
-    .await
-    .map_err(|e| AdminUiError::Internal(format!("folder-templates task: {e}")))?
-}
-
-/// One named console-local folder template: a display name (the key) plus the
-/// canonical FOLDER JSON tree it commits. No openEHR spec governs an admin-UI
-/// convenience like this — our own design/extension; the FOLDER shape it
-/// carries IS spec-bound (ITS-REST `specifications/schemas/ehr/Folder.yaml`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FolderTemplate {
-    /// Display name (also the template key).
-    pub name: String,
-    /// The canonical FOLDER JSON tree (root FOLDER, no `uid`).
-    pub folder: Value,
-}
-
-/// The standard openEHR directory root archetype id (the FOLDER example in
-/// ITS-REST `specifications/schemas/ehr/Folder.yaml`).
-pub(crate) const DIRECTORY_ARCHETYPE: &str = "openEHR-EHR-FOLDER.directory.v1";
-
-/// Node id reused by child folders; siblings differ by `name` (RM common
-/// `master05-directory_package` §Paths — uniqueness modifiers). Consumed
-/// only by the ssr-side folder-template builder, so it is feature-gated
-/// (the wasm lib would flag it dead).
-#[cfg(feature = "ssr")]
-pub(crate) const FOLDER_NODE_ID: &str = "at0001";
-
-/// Build a canonical FOLDER JSON node from `archetype_node_id`, `name`, and
-/// child `folders`. `_type` + mandatory `archetype_node_id`/`name` (from
-/// LOCATABLE) + the FOLDER `folders`/`items` arrays per
-/// `specifications/schemas/ehr/Folder.yaml`; no `uid` — the CDR assigns the
-/// `OBJECT_VERSION_ID` on create.
-pub(crate) fn folder_json(archetype_node_id: &str, name: &str, folders: &[Value]) -> Value {
-    json!({
-        "_type": "FOLDER",
-        "archetype_node_id": archetype_node_id,
-        "name": { "_type": "DV_TEXT", "value": name },
-        "folders": folders,
-        "items": [],
-    })
-}
-
-/// An empty directory root FOLDER (the "empty root" create choice).
-pub(crate) fn empty_root_folder() -> Value {
-    folder_json(DIRECTORY_ARCHETYPE, "root", &[])
-}
-
-/// The toast detail for a directory write: the new version, or a generic line
-/// when the CDR returned no representation body.
-fn directory_toast_detail(uid: &str) -> String {
-    if uid.is_empty() {
-        "The directory was committed.".to_owned()
-    } else {
-        format!("New version {uid}")
-    }
-}
-
-/// List an EHR's compositions via [`LIST_COMPOSITIONS_AQL`], one page at
-/// `offset`.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session; CDR transport
-/// errors pass through; a non-2xx CDR answer normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success);
-/// [`AdminUiError::Internal`] when the result set is not valid JSON.
-#[server]
-pub async fn list_compositions(ehr_id: String, offset: u32) -> Result<ResultPage, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let url = state.cdr.rest_v1("query/aql");
-    let body = aql_request_body(
-        LIST_COMPOSITIONS_AQL,
-        &serde_json::json!({ "ehr_id": ehr_id }),
-        offset,
-    );
-    let response = state
-        .cdr
-        .post(
-            &session.credential,
-            &url,
-            "application/json",
-            "application/json",
-            &[],
-            body,
-        )
-        .await?;
-    let response = crate::cdr::CdrClient::expect_success(response)?;
-    parse_result_set(&response.body, offset)
-}
-
-#[cfg(feature = "ssr")]
-/// The new version uid of a just-committed COMPOSITION: `uid.value` from the
-/// `Prefer: return=representation` body (an `OBJECT_VERSION_ID`). Empty when
-/// the CDR returned no representation body — the UI then shows a generic
-/// success message rather than a uid.
-pub(crate) fn commit_version_uid(body: &str) -> String {
-    serde_json::from_str::<Value>(body)
-        .ok()
-        .and_then(|doc| {
-            doc.get("uid")
-                .and_then(|u| u.get("value"))
-                .and_then(Value::as_str)
-                .map(str::to_owned)
-        })
-        .unwrap_or_default()
-}
-
-/// Commit a new COMPOSITION to the EHR (`POST /ehr/{ehr_id}/composition`). The
-/// `format` picks the `Content-Type` (canonical JSON `application/json`,
-/// canonical XML `application/xml`, FLAT `application/openehr.wt.flat+json`);
-/// a FLAT commit additionally requires the `openehr-template-id` header.
-/// `Accept: application/json` + `Prefer: return=representation` yields a
-/// canonical composition body whose `uid.value` is returned as the new
-/// version uid.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session;
-/// [`AdminUiError::Invalid`] on an empty body or a FLAT commit without a
-/// template id; CDR transport errors pass through; a non-2xx CDR answer (its
-/// validation diagnostics, which the UI renders verbatim, included)
-/// normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn commit_composition(
-    ehr_id: String,
-    format: ReprFormat,
-    template_id: String,
-    body: String,
-) -> Result<String, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    if body.trim().is_empty() {
-        return Err(AdminUiError::Invalid(
-            "the composition body is empty".to_owned(),
-        ));
-    }
-    let template_id = template_id.trim();
-    let mut headers: Vec<(&str, &str)> = vec![("Prefer", "return=representation")];
-    if matches!(format, ReprFormat::Flat) {
-        if template_id.is_empty() {
-            return Err(AdminUiError::Invalid(
-                "a template id is required to commit a FLAT composition".to_owned(),
-            ));
-        }
-        headers.push(("openehr-template-id", template_id));
-    }
-    let url = state
-        .cdr
-        .rest_v1(&format!("ehr/{}/composition", urlencoding::encode(&ehr_id)));
-    let response = state
-        .cdr
-        .post(
-            &session.credential,
-            &url,
-            format.media_type(),
-            "application/json",
-            &headers,
-            body,
-        )
-        .await?;
-    let response = crate::cdr::CdrClient::expect_success(response)?;
-    Ok(commit_version_uid(&response.body))
-}
-
-/// Rows fetched per page of the contribution list (fixed, per the tab's
-/// prev/next paging).
-const CONTRIBUTION_FETCH: u32 = 20;
-
-/// One row of the EHR's contribution list. Fixed-size-safe strings (rules §1);
-/// the shape is the CDR's contribution-list contract
-/// (`{uid, time_committed, committer, change_type}`).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ContributionRow {
-    /// The CONTRIBUTION uid.
-    pub uid: String,
-    /// `AUDIT_DETAILS.time_committed` value.
-    pub time_committed: String,
-    /// `AUDIT_DETAILS.committer` name.
-    pub committer: String,
-    /// `AUDIT_DETAILS.change_type` value.
-    pub change_type: String,
-}
-
-/// One page of an EHR's contributions: the rows plus the total count (for
-/// prev/next). Carries only fixed-size ints so it is WASM-safe over the
-/// server-fn boundary (rules §1).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ContributionPage {
-    /// The contribution rows on this page.
-    pub rows: Vec<ContributionRow>,
-    /// The total number of contributions in the EHR.
-    pub total: u32,
-}
-
-/// List an EHR's contributions, one page at `offset`
-/// (`GET /ehr/{ehr_id}/contribution?offset&fetch`). `fetch` is fixed at
-/// [`CONTRIBUTION_FETCH`].
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session; CDR transport
-/// errors pass through; a non-2xx CDR answer normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success) — a
-/// `404`/`405` from an older CDR that lacks the list route surfaces there as
-/// [`AdminUiError::Cdr`] (the tab renders it inline);
-/// [`AdminUiError::Internal`] when the page is not valid JSON.
-#[server]
-pub async fn list_contributions(
-    ehr_id: String,
-    offset: u32,
-) -> Result<ContributionPage, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let url = state.cdr.rest_v1(&format!(
-        "ehr/{}/contribution?offset={offset}&fetch={CONTRIBUTION_FETCH}",
-        urlencoding::encode(&ehr_id),
-    ));
-    let response = state
-        .cdr
-        .get(&session.credential, &url, "application/json")
-        .await?;
-    let response = crate::cdr::CdrClient::expect_success(response)?;
-    parse_contributions(&response.body)
-}
-
-#[cfg(feature = "ssr")]
-/// Parse the contribution-list body (`{ "rows": [...], "total": N }`) into a
-/// [`ContributionPage`]. Defensive throughout — a missing field reads as empty
-/// / zero rather than failing.
-///
-/// # Errors
-/// [`AdminUiError::Internal`] when the body is not valid JSON.
-pub(crate) fn parse_contributions(body: &str) -> Result<ContributionPage, AdminUiError> {
-    let doc: Value = serde_json::from_str(body)
-        .map_err(|e| AdminUiError::Internal(format!("contributions JSON: {e}")))?;
-    let rows = doc
-        .get("rows")
-        .and_then(Value::as_array)
-        .map(|rows| rows.iter().map(contribution_row).collect())
-        .unwrap_or_default();
-    let total = doc
-        .get("total")
-        .and_then(Value::as_u64)
-        .and_then(|n| u32::try_from(n).ok())
-        .unwrap_or_default();
-    Ok(ContributionPage { rows, total })
-}
-
-#[cfg(feature = "ssr")]
-/// Flatten one contribution-list entry into a [`ContributionRow`], each field
-/// read defensively as a string.
-fn contribution_row(value: &Value) -> ContributionRow {
-    ContributionRow {
-        uid: contribution_field(value, "uid"),
-        time_committed: contribution_field(value, "time_committed"),
-        committer: contribution_field(value, "committer"),
-        change_type: contribution_field(value, "change_type"),
-    }
-}
-
-#[cfg(feature = "ssr")]
-/// Read a top-level string field, or an empty string when absent / not a string.
-fn contribution_field(value: &Value, key: &str) -> String {
-    value
-        .get(key)
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-        .to_owned()
-}
-
-/// Look up a single CONTRIBUTION by uid
-/// (`GET /ehr/{ehr_id}/contribution/{contribution_uid}`) — the by-uid lookup
-/// box the Contributions tab keeps below its list (see
-/// [`contributions_section`]). Returns the raw canonical JSON.
-///
-/// # Errors
-/// [`AdminUiError::Unauthenticated`] without a console session; CDR transport
-/// errors pass through; a non-2xx CDR answer (a `404` for an unknown uid
-/// included) normalizes via
-/// [`CdrClient::expect_success`](crate::cdr::CdrClient::expect_success).
-#[server]
-pub async fn fetch_contribution(
-    ehr_id: String,
-    contribution_uid: String,
-) -> Result<String, AdminUiError> {
-    let session = crate::session::require_session().await?;
-    let state: crate::state::AppState = leptos::prelude::expect_context();
-    let url = state.cdr.rest_v1(&format!(
-        "ehr/{}/contribution/{}",
-        urlencoding::encode(&ehr_id),
-        urlencoding::encode(&contribution_uid)
-    ));
-    let response = state
-        .cdr
-        .get(&session.credential, &url, "application/json")
-        .await?;
-    Ok(crate::cdr::CdrClient::expect_success(response)?.body)
-}
-
-/// The `/ehrs/{ehr_id}` screen: the tab bar plus four always-mounted,
-/// visibility-toggled tab bodies.
-#[allow(clippy::must_use_candidate)] // #[component] rewrites the fn; view!/mount always consumes the value
-#[component]
-pub fn EhrDetailPage() -> impl IntoView {
-    let params = leptos_router::hooks::use_params_map();
-    let ehr_id = Signal::derive(move || params.with(|p| p.get("ehr_id").unwrap_or_default()));
-    let query = leptos_router::hooks::use_query_map();
-    let offset = Signal::derive(move || {
-        query
-            .with(|q| q.get("offset"))
-            .and_then(|s| s.parse::<u32>().ok())
-            .unwrap_or(0)
-    });
-    // Tab state lives in the URL (`?tab=`, rules §9): shareable and refresh-safe.
-    // A Memo (not an Effect) derives the active tab, defaulting to "status".
-    let selected: Memo<String> = Memo::new(move |_| {
-        query
-            .with(|q| q.get("tab"))
-            .filter(|t| !t.is_empty())
-            .unwrap_or_else(|| "status".to_owned())
-    });
-
-    let status = status_section(ehr_id, selected);
-    let directory = directory_section(ehr_id, selected);
-    let compositions = compositions_section(ehr_id, offset, selected);
-    let contributions = contributions_section(ehr_id, selected);
-
-    let heading = Signal::derive(move || {
-        let id = ehr_id.get();
-        let short: String = id.chars().take(8).collect();
-        format!("EHR {short}…")
-    });
-
-    let tabs = tab_bar(ehr_id, selected);
-
-    view! {
-        <Title text="EHR detail · ehrbase-admin" />
-        <div class="p-6">
-            <PageHeader
-                title=Signal::derive(move || heading.get())
-                crumbs=vec![Crumb::new("EHRs", "/ehrs")]
-                mono=true
-            />
-            {tabs}
-            <div class="mt-4">
-                <div class:hidden=move || selected.get() != "status">{status}</div>
-                <div class:hidden=move || selected.get() != "directory">{directory}</div>
-                <div class:hidden=move || selected.get() != "compositions">{compositions}</div>
-                <div class:hidden=move || {
-                    selected.get() != "contributions"
-                }>{contributions}</div>
-            </div>
-        </div>
-    }
-}
-
-/// The URL-driven tab bar: four pill anchors (`?tab=…`) replacing the thaw
-/// `TabList`. Selected = `bg-accent-subtle text-accent-ink`; idle =
-/// `text-ink-muted hover:bg-sunken`. Plain anchors keep the tabs working
-/// before hydration (the router intercepts them once WASM loads).
-fn tab_bar(ehr_id: Signal<String>, selected: Memo<String>) -> AnyView {
-    let link = move |value: &'static str, label: &'static str| {
-        let href = move || format!("/ehrs/{}?tab={value}", ehr_id.get());
-        let class = move || {
-            if selected.get() == value {
-                "rounded-control px-3 py-1.5 text-sm font-medium bg-accent-subtle text-accent-ink"
-            } else {
-                "rounded-control px-3 py-1.5 text-sm font-medium text-ink-muted hover:bg-sunken"
-            }
-        };
-        view! {
-            <a href=href class=class>
-                {label}
-            </a>
-        }
-    };
-    view! {
-        <div class="flex flex-wrap gap-1 border-b border-edge pb-2">
-            {link("status", "Status")} {link("directory", "Directory")}
-            {link("compositions", "Compositions")} {link("contributions", "Contributions")}
-        </div>
-    }
-    .into_any()
-}
-
-/// Status tab: `fetch_ehr_status` → queryable/modifiable badges, the subject,
-/// and the raw JSON in a [`DocumentPane`]. The source is gated on the tab
-/// being active so it fetches only when shown.
-fn status_section(ehr_id: Signal<String>, selected: Memo<String>) -> AnyView {
-    let resource = Resource::new(
-        move || (selected.get() == "status").then(|| ehr_id.get()),
-        |active| async move {
-            match active {
-                Some(id) => fetch_ehr_status(id).await.map(Some),
-                None => Ok(None),
-            }
-        },
-    );
-    view! {
-        <Suspense fallback=table_skeleton>
-            {move || Suspend::new(async move {
-                let rendered = resource
-                    .await
-                    .and_then(|opt| match opt {
-                        Some(body) => status_body(&body),
-                        None => Ok(().into_any()),
-                    });
-                match rendered {
-                    Ok(view) => view,
-                    Err(e) => crate::components::format_view::inline_error(&e),
-                }
-            })}
-        </Suspense>
-    }
-    .into_any()
-}
-
-/// Render an `EHR_STATUS` document: the two capability badges, the subject
-/// reference, and the raw JSON.
-///
-/// # Errors
-/// [`AdminUiError::Internal`] when the body is not valid JSON.
-fn status_body(body: &str) -> Result<AnyView, AdminUiError> {
-    let doc: Value = serde_json::from_str(body)
-        .map_err(|e| AdminUiError::Internal(format!("ehr_status JSON: {e}")))?;
-    let queryable = doc
-        .get("is_queryable")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let modifiable = doc
-        .get("is_modifiable")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let subject = doc
-        .get("subject")
-        .and_then(|s| s.get("external_ref"))
-        .and_then(|r| r.get("id"))
-        .and_then(|i| i.get("value"))
-        .and_then(Value::as_str)
-        .map_or_else(|| "self (no external subject)".to_owned(), str::to_owned);
-    let pretty =
-        crate::components::format_view::pretty_body(body, crate::format::ReprFormat::CanonicalJson);
-    let doc_sig = RwSignal::new(pretty);
-    Ok(view! {
-        <div class=format!("{CARD_PAD} flex flex-col gap-3")>
-            <div class="flex flex-wrap gap-2 items-center">
-                {capability_badge("queryable", queryable)}
-                {capability_badge("modifiable", modifiable)}
-            </div>
-            <div class="text-sm">
-                <span class="font-medium text-ink-muted">"subject: "</span>
-                <span class="font-mono break-all text-ink">{subject}</span>
-            </div>
-            {(!queryable)
-                .then(|| {
-                    view! {
-                        <div
-                            role="status"
-                            class="rounded-control border border-warn/40 bg-warn-subtle px-3 py-2 text-sm text-warn"
-                        >
-                            "This EHR is not queryable — AQL over it returns nothing."
-                        </div>
-                    }
-                })}
-            <DocumentPane body=doc_sig />
-        </div>
-    }
-    .into_any())
-}
-
-/// An ok/danger capability chip for an `EHR_STATUS` boolean flag.
-fn capability_badge(label: &'static str, on: bool) -> AnyView {
-    let (mark, class) = if on {
-        ("✓", "bg-ok-subtle text-ok")
-    } else {
-        ("✗", "bg-danger-subtle text-danger")
-    };
-    view! {
-        <span class=format!(
-            "inline-flex items-center gap-1 rounded-control px-2 py-0.5 text-xs font-medium {class}",
-        )>{mark} " " {label}</span>
-    }
-    .into_any()
-}
-
-/// Directory tab: `fetch_directory` → the recursive `FOLDER` tree with an
-/// "Edit as JSON" editor committing a `PUT` update, or — when the EHR has no
-/// directory (the CDR 404s) — a "Create directory" section that commits a
-/// `POST` from a folder template or an empty root. The list resource depends
-/// on both write actions' versions, so a successful commit refetches it (rules
-/// §6 — never fetch-in-effect).
-fn directory_section(ehr_id: Signal<String>, selected: Memo<String>) -> AnyView {
-    let toaster = thaw::ToasterInjection::expect_context();
-    let create = Action::new(|(ehr_id, body): &(String, String)| {
-        let ehr_id = ehr_id.clone();
-        let body = body.clone();
-        async move { create_directory(ehr_id, body).await }
-    });
-    let update = Action::new(|(ehr_id, uid, body): &(String, String, String)| {
-        let ehr_id = ehr_id.clone();
-        let uid = uid.clone();
-        let body = body.clone();
-        async move { update_directory(ehr_id, uid, body).await }
-    });
-    // Toast either write's success (an outside-world side-effect — rules §2);
-    // the failure diagnostic stays inline in the WELL feedback pane.
-    Effect::new(move |_| {
-        if let Some(Ok(uid)) = create.value().get() {
-            toast_success(toaster, "Directory created", &directory_toast_detail(&uid));
-        }
-    });
-    Effect::new(move |_| {
-        if let Some(Ok(uid)) = update.value().get() {
-            toast_success(toaster, "Directory updated", &directory_toast_detail(&uid));
-        }
-    });
-    // Both resources are created ONCE here (never inside the Suspend — rules
-    // §4); the templates load feeds only the create section.
-    let directory = Resource::new(
-        move || {
-            let versions = (create.version().get(), update.version().get());
-            (selected.get() == "directory").then(|| (ehr_id.get(), versions))
-        },
-        |active| async move {
-            match active {
-                Some((id, _)) => fetch_directory(id).await,
-                None => Ok(None),
-            }
-        },
-    );
-    let templates = Resource::new(
-        move || (selected.get() == "directory").then_some(()),
-        |active| async move {
-            match active {
-                Some(()) => list_folder_templates().await.map(Some),
-                None => Ok(None),
-            }
-        },
-    );
-    view! {
-        <Suspense fallback=table_skeleton>
-            {move || Suspend::new(async move {
-                match directory.await {
-                    Ok(Some(state)) => directory_edit_section(&state, ehr_id, update),
-                    Ok(None) => {
-                        match templates.await {
-                            Ok(list) => {
-                                directory_create_section(list.unwrap_or_default(), ehr_id, create)
-                            }
-                            Err(e) => crate::components::format_view::inline_error(&e),
-                        }
-                    }
-                    Err(e) => crate::components::format_view::inline_error(&e),
-                }
-            })}
-        </Suspense>
-    }
-    .into_any()
-}
-
-/// The existing-directory view: the recursive `FOLDER` tree plus an
-/// "Edit as JSON" toggle whose textarea (prefilled with the current canonical
-/// FOLDER JSON) commits a `PUT` update via the shared `update` action (mirrors
-/// the composition commit editor). The editor block stays mounted and is
-/// toggled with `class:hidden` (rules §8 — identical server/client structure).
-fn directory_edit_section(
-    state: &DirectoryState,
-    ehr_id: Signal<String>,
-    update: Action<(String, String, String), Result<String, AdminUiError>>,
-) -> AnyView {
-    let doc: Value = match serde_json::from_str(&state.body) {
-        Ok(value) => value,
-        Err(e) => {
-            return crate::components::format_view::inline_error(&AdminUiError::Internal(format!(
-                "directory JSON: {e}"
-            )));
-        }
-    };
-    let tree = folder_node(&doc);
-    let version_uid = state.version_uid.clone();
-    let pretty = crate::components::format_view::pretty_body(
-        &state.body,
-        crate::format::ReprFormat::CanonicalJson,
-    );
-    let editing = RwSignal::new(false);
-    let body = RwSignal::new(pretty);
-    let on_commit = move |_| {
-        update.dispatch((ehr_id.get(), version_uid.clone(), body.get()));
-    };
-    view! {
-        <section class=CARD_PAD>
-            <div class="flex flex-wrap items-center justify-between gap-2">
-                <h2 class=CARD_TITLE>"Directory"</h2>
-                <button
-                    id="directory-edit"
-                    type="button"
-                    class=BTN_SECONDARY
-                    on:click=move |_| editing.update(|open| *open = !*open)
-                >
-                    {move || if editing.get() { "Hide editor" } else { "Edit as JSON" }}
-                </button>
-            </div>
-            <ul class="text-sm text-ink">{tree}</ul>
-            <div class="mt-3 flex flex-col gap-3" class:hidden=move || !editing.get()>
-                <div class="flex flex-col gap-1">
-                    <label class=LABEL r#for="directory-body">
-                        "Directory FOLDER (canonical JSON)"
-                    </label>
-                    <textarea
-                        id="directory-body"
-                        class=format!("{TEXTAREA} min-h-[16rem]")
-                        prop:value=move || body.get()
-                        on:input:target=move |ev| body.set(ev.target().value())
-                    >
-                        {body.get_untracked()}
-                    </textarea>
-                </div>
-                <div class="flex items-center gap-3">
-                    <button
-                        id="directory-commit"
-                        type="button"
-                        class=BTN_PRIMARY
-                        disabled=Signal::derive(move || update.pending().get())
-                        on:click=on_commit
-                    >
-                        "Save directory"
-                    </button>
-                    <Show when=move || update.pending().get()>
-                        <span class="text-sm text-ink-muted">"Saving…"</span>
-                    </Show>
-                </div>
-                {directory_feedback(update)}
-            </div>
-        </section>
-    }
-    .into_any()
-}
-
-/// The no-directory view: a "Create directory" section choosing a folder
-/// template (or an empty root), previewing the FOLDER tree to be committed,
-/// and committing a `POST` create via the shared `create` action.
-fn directory_create_section(
-    templates: Vec<FolderTemplate>,
-    ehr_id: Signal<String>,
-    create: Action<(String, String), Result<String, AdminUiError>>,
-) -> AnyView {
-    // The empty string selects the empty root; any other value names a template.
-    let choice = RwSignal::new(String::new());
-    let templates_for_pick = templates.clone();
-    let chosen = Signal::derive(move || {
-        let key = choice.get();
-        if key.is_empty() {
-            empty_root_folder()
-        } else {
-            templates_for_pick
-                .iter()
-                .find(|t| t.name == key)
-                .map_or_else(empty_root_folder, |t| t.folder.clone())
-        }
-    });
-    let on_create = move |_| {
-        let body = serde_json::to_string(&chosen.get()).unwrap_or_default();
-        create.dispatch((ehr_id.get(), body));
-    };
-    let options = templates
-        .into_iter()
-        .map(|template| {
-            let value = template.name.clone();
-            let label = template.name;
-            view! { <option value=value>{label}</option> }.into_any()
-        })
-        .collect::<Vec<_>>();
-    view! {
-        <section class=CARD_PAD>
-            <h2 class=CARD_TITLE>"Create directory"</h2>
-            <p class="mb-3 text-sm text-ink-muted">
-                "This EHR has no directory yet. Start from a folder template or an empty root."
-            </p>
-            <div class="flex flex-col gap-3">
-                <div class="flex flex-col gap-1">
-                    <label class=LABEL r#for="folder-template">
-                        "Folder template"
-                    </label>
-                    <select
-                        id="folder-template"
-                        class=SELECT
-                        prop:value=move || choice.get()
-                        on:change=move |ev| choice.set(event_target_value(&ev))
-                    >
-                        <option value="">"Empty root"</option>
-                        {options}
-                    </select>
-                </div>
-                <div class="flex flex-col gap-1">
-                    <span class=LABEL>"Preview"</span>
-                    <div class=WELL>
-                        <ul class="text-sm text-ink">{move || folder_node(&chosen.get())}</ul>
-                    </div>
-                </div>
-                <div class="flex items-center gap-3">
-                    <button
-                        id="directory-create"
-                        type="button"
-                        class=BTN_PRIMARY
-                        disabled=Signal::derive(move || create.pending().get())
-                        on:click=on_create
-                    >
-                        "Create directory"
-                    </button>
-                    <Show when=move || create.pending().get()>
-                        <span class="text-sm text-ink-muted">"Creating…"</span>
-                    </Show>
-                </div>
-                {directory_feedback(create)}
-            </div>
-        </section>
-    }
-    .into_any()
-}
-
-/// A directory write action's failure pane: the CDR's validation diagnostics
-/// verbatim in a scrollable WELL (a `<pre>`, mirroring the composition commit
-/// feedback). Success is a toast (see [`directory_section`]).
-fn directory_feedback<I: Send + Sync + 'static>(
-    action: Action<I, Result<String, AdminUiError>>,
-) -> AnyView {
-    view! {
-        {move || match action.value().get() {
-            Some(Err(error)) => {
-                view! {
-                    <div class=WELL>
-                        <pre class="overflow-auto max-h-[40vh] whitespace-pre-wrap font-mono text-xs text-danger">
-                            {error.to_string()}
-                        </pre>
-                    </div>
-                }
-                    .into_any()
-            }
-            _ => ().into_any(),
-        }}
-    }
-    .into_any()
-}
-
-/// One `FOLDER` node: its name, its child folders (recursively), and its
-/// item references. Returns [`AnyView`] (recursive tree fns erase — rules §1).
-fn folder_node(folder: &Value) -> AnyView {
-    let name = folder
-        .get("name")
-        .and_then(|n| n.get("value"))
-        .and_then(Value::as_str)
-        .unwrap_or("(folder)")
-        .to_owned();
-    let subfolders = folder
-        .get("folders")
-        .and_then(Value::as_array)
-        .map(|folders| folders.iter().map(folder_node).collect::<Vec<_>>())
-        .unwrap_or_default();
-    let items = folder
-        .get("items")
-        .and_then(Value::as_array)
-        .map(|items| items.iter().map(item_ref_node).collect::<Vec<_>>())
-        .unwrap_or_default();
-    view! {
-        <li class="py-0.5">
-            <span class="font-medium text-ink">"📁 " {name}</span>
-            <ul class="pl-4 ml-2 border-l border-edge">{subfolders} {items}</ul>
-        </li>
-    }
-    .into_any()
-}
-
-/// One `OBJECT_REF` item under a folder: its type and id value.
-fn item_ref_node(item: &Value) -> AnyView {
-    let id = item
-        .get("id")
-        .and_then(|i| i.get("value"))
-        .and_then(Value::as_str)
-        .unwrap_or("(ref)")
-        .to_owned();
-    let ref_type = item
-        .get("type")
-        .and_then(Value::as_str)
-        .unwrap_or("OBJECT")
-        .to_owned();
-    view! {
-        <li class="py-0.5 text-ink-muted">
-            "• " <span class="uppercase text-xs mr-1">{ref_type}</span>
-            <span class="font-mono break-all">{id}</span>
-        </li>
-    }
-    .into_any()
-}
-
-/// Compositions tab: `list_compositions` (AQL) → a paged table whose uid
-/// cells link to the composition viewer (under `<Transition>` so paging keeps
-/// old rows visible), plus a "Commit composition" form below it. A successful
-/// commit bumps the commit action's version — a source of the list resource —
-/// refetching the table (rules §6 — never fetch-in-effect).
-fn compositions_section(
-    ehr_id: Signal<String>,
-    offset: Signal<u32>,
-    selected: Memo<String>,
-) -> AnyView {
-    let toaster = thaw::ToasterInjection::expect_context();
-    let commit = Action::new(
-        |(ehr_id, format, template_id, body): &(String, ReprFormat, String, String)| {
-            let ehr_id = ehr_id.clone();
-            let format = *format;
-            let template_id = template_id.clone();
-            let body = body.clone();
-            async move { commit_composition(ehr_id, format, template_id, body).await }
-        },
-    );
-    // Toast the outcome on success (an outside-world side-effect — rules §2);
-    // the failure diagnostic stays inline in the form.
-    Effect::new(move |_| {
-        if let Some(Ok(uid)) = commit.value().get() {
-            let detail = if uid.is_empty() {
-                "The composition was committed.".to_owned()
-            } else {
-                format!("New version {uid}")
-            };
-            toast_success(toaster, "Composition committed", &detail);
-        }
-    });
-    let resource = Resource::new(
-        move || {
-            let version = commit.version().get();
-            (selected.get() == "compositions").then(|| (ehr_id.get(), offset.get(), version))
-        },
-        |active| async move {
-            match active {
-                Some((id, offset, _)) => list_compositions(id, offset).await.map(Some),
-                None => Ok(None),
-            }
-        },
-    );
-    let table = view! {
-        <Transition fallback=table_skeleton>
-            {move || Suspend::new(async move {
-                match resource.await {
-                    Ok(Some(page)) => compositions_table(&page, &ehr_id.get()),
-                    Ok(None) => ().into_any(),
-                    Err(e) => crate::components::format_view::inline_error(&e),
-                }
-            })}
-        </Transition>
-    }
-    .into_any();
-    let form = commit_form(ehr_id, commit);
-    view! { <div>{table} {form}</div> }.into_any()
-}
-
-/// The "Commit composition" form: a format select, a template-id input shown
-/// only for FLAT (its `openehr-template-id` header is required there — kept in
-/// the DOM and toggled with `class:hidden` so the server and client view
-/// structure stay identical, rules §8), a large body textarea, and a Commit
-/// button dispatching the shared `commit` action.
-fn commit_form(
-    ehr_id: Signal<String>,
-    commit: Action<(String, ReprFormat, String, String), Result<String, AdminUiError>>,
-) -> AnyView {
-    let format = RwSignal::new(ReprFormat::CanonicalJson);
-    let template_id = RwSignal::new(String::new());
-    let body = RwSignal::new(String::new());
-    let is_flat = move || format.get() == ReprFormat::Flat;
-    let on_commit = move |_| {
-        commit.dispatch((ehr_id.get(), format.get(), template_id.get(), body.get()));
-    };
-    view! {
-        <section class=format!("{CARD_PAD} mt-4")>
-            <h2 class=CARD_TITLE>"Commit composition"</h2>
-            <div class="flex flex-col gap-3">
-                <div class="flex flex-wrap items-end gap-3">
-                    <div class="flex flex-col gap-1">
-                        <label class=LABEL r#for="commit-format">
-                            "Format"
-                        </label>
-                        <select
-                            id="commit-format"
-                            class=SELECT
-                            prop:value=move || format_value(format.get())
-                            on:change=move |ev| {
-                                format.set(format_from_value(&event_target_value(&ev)));
-                            }
-                        >
-                            <option value="json">"Canonical JSON"</option>
-                            <option value="xml">"Canonical XML"</option>
-                            <option value="flat">"FLAT"</option>
-                        </select>
-                    </div>
-                    <div class="flex flex-col gap-1" class:hidden=move || !is_flat()>
-                        <label class=LABEL r#for="commit-template-id">
-                            "Template id"
-                        </label>
-                        <input
-                            id="commit-template-id"
-                            type="text"
-                            class=INPUT
-                            placeholder="template id (required for FLAT)"
-                            prop:value=move || template_id.get()
-                            on:input:target=move |ev| template_id.set(ev.target().value())
-                        />
-                    </div>
-                </div>
-                <div class="flex flex-col gap-1">
-                    <label class=LABEL r#for="commit-body">
-                        "Composition"
-                    </label>
-                    <textarea
-                        id="commit-body"
-                        class=format!("{TEXTAREA} min-h-[16rem]")
-                        placeholder="paste the composition document (JSON, XML, or FLAT)…"
-                        prop:value=move || body.get()
-                        on:input:target=move |ev| body.set(ev.target().value())
-                    >
-                        {body.get_untracked()}
-                    </textarea>
-                </div>
-                <div class="flex items-center gap-3">
-                    <button
-                        id="commit-submit"
-                        type="button"
-                        class=BTN_PRIMARY
-                        disabled=Signal::derive(move || commit.pending().get())
-                        on:click=on_commit
-                    >
-                        "Commit"
-                    </button>
-                    <Show when=move || commit.pending().get()>
-                        <span class="text-sm text-ink-muted">"Committing…"</span>
-                    </Show>
-                </div>
-                {commit_feedback(commit)}
-            </div>
-        </section>
-    }
-    .into_any()
-}
-
-/// The commit action's failure pane: the CDR's validation diagnostics
-/// verbatim in a scrollable WELL (they are long and precious — a `<pre>`, not
-/// a one-line error). Success is a toast (see [`compositions_section`]).
-fn commit_feedback(
-    commit: Action<(String, ReprFormat, String, String), Result<String, AdminUiError>>,
-) -> AnyView {
-    view! {
-        {move || match commit.value().get() {
-            Some(Err(error)) => {
-                view! {
-                    <div class=WELL>
-                        <pre class="overflow-auto max-h-[40vh] whitespace-pre-wrap font-mono text-xs text-danger">
-                            {error.to_string()}
-                        </pre>
-                    </div>
-                }
-                    .into_any()
-            }
-            _ => ().into_any(),
-        }}
-    }
-    .into_any()
-}
-
-/// The `<select>` option value for a committable format.
-fn format_value(format: ReprFormat) -> &'static str {
-    match format {
-        ReprFormat::CanonicalXml => "xml",
-        ReprFormat::Flat => "flat",
-        _ => "json",
-    }
-}
-
-/// The committable format for a `<select>` option value (unknown → canonical
-/// JSON).
-fn format_from_value(value: &str) -> ReprFormat {
-    match value {
-        "xml" => ReprFormat::CanonicalXml,
-        "flat" => ReprFormat::Flat,
-        _ => ReprFormat::CanonicalJson,
-    }
-}
-
-/// Render one page of compositions: a table whose uid cell links to the
-/// composition viewer (the versioned-object id — any `::system::version`
-/// suffix stripped for the link, the full uid kept visible), plus paging.
-fn compositions_table(page: &ResultPage, ehr_id: &str) -> AnyView {
-    if page.rows.is_empty() {
-        return view! { <p class="text-sm text-ink-muted">"No compositions in this EHR."</p> }
-            .into_any();
-    }
-    let rows = page.rows.clone();
-    let ehr_id_owned = ehr_id.to_owned();
-    let body = view! {
-        <For
-            each=move || rows.clone()
-            key=|row| row.first().map(cell_text).unwrap_or_default()
-            let:row
-        >
-            {composition_row(&row, &ehr_id_owned)}
-        </For>
-    }
-    .into_any();
-    let paging = paging_controls(page.offset, page.rows.len(), &format!("/ehrs/{ehr_id}"));
-    view! {
-        {table_shell(&["Composition", "Name", "Template", "Started"], body)}
-        {paging}
-    }
-    .into_any()
-}
-
-/// One composition row: the uid cell links to the viewer at the
-/// versioned-object id; the full uid stays visible.
-fn composition_row(row: &[Value], ehr_id: &str) -> AnyView {
-    let uid = row.first().map(cell_text).unwrap_or_default();
-    let vo_id = versioned_object_id(&uid).to_owned();
-    let cells = row
-        .iter()
-        .enumerate()
-        .map(|(i, value)| {
-            let text = cell_text(value);
-            if i == 0 {
-                let href = format!("/ehrs/{ehr_id}/compositions/{vo_id}");
-                view! {
-                    <td class=CELL_MONO>
-                        <A href=href attr:class="text-accent hover:underline">
-                            {text}
-                        </A>
-                    </td>
-                }
-                .into_any()
-            } else {
-                view! { <td class=CELL>{text}</td> }.into_any()
-            }
-        })
-        .collect::<Vec<_>>();
-    view! { <tr class=ROW>{cells}</tr> }.into_any()
-}
-
-/// The versioned-object id from an `OBJECT_VERSION_ID` value: everything
-/// before the first `::` (`uuid::system::version` → `uuid`), which is what
-/// the composition route keys on.
-fn versioned_object_id(uid: &str) -> &str {
-    uid.split_once("::").map_or(uid, |(head, _)| head)
-}
-
-/// Contributions tab: a paged list of the EHR's contributions
-/// (`list_contributions` → `GET /ehr/{ehr_id}/contribution?offset&fetch`) under
-/// a `<Transition>`, plus the by-uid lookup box kept below it. An older CDR
-/// that lacks the list route (a `404`/`405`) renders inline via the normal
-/// error path — the lookup box still works.
-fn contributions_section(ehr_id: Signal<String>, selected: Memo<String>) -> AnyView {
-    // Paging state is a local signal — the tab itself is already URL-state.
-    let offset = RwSignal::new(0_u32);
-    let list = Resource::new(
-        move || (selected.get() == "contributions").then(|| (ehr_id.get(), offset.get())),
-        |active| async move {
-            match active {
-                Some((id, off)) => list_contributions(id, off).await.map(Some),
-                None => Ok(None),
-            }
-        },
-    );
-    let table = view! {
-        <Transition fallback=table_skeleton>
-            {move || Suspend::new(async move {
-                match list.await {
-                    Ok(Some(page)) => contributions_table(&page, offset),
-                    Ok(None) => ().into_any(),
-                    Err(e) => crate::components::format_view::inline_error(&e),
-                }
-            })}
-        </Transition>
-    }
-    .into_any();
-    let lookup = contribution_lookup(ehr_id, selected);
-    view! { <div class="flex flex-col gap-4">{table} {lookup}</div> }.into_any()
-}
-
-/// Render one page of contributions as the shared [`table_shell`] plus
-/// prev/next paging (local-signal driven). An empty page is an
-/// [`EmptyState`], not bare muted text.
-fn contributions_table(page: &ContributionPage, offset: RwSignal<u32>) -> AnyView {
-    if page.rows.is_empty() {
-        return view! {
-            <EmptyState
-                icon=icondata_lu::LuInbox
-                message="No contributions"
-                hint="This EHR has no contributions on this page."
-            />
-        }
-        .into_any();
-    }
-    let rows = page.rows.clone();
-    let body = view! {
-        <For each=move || rows.clone() key=|row| row.uid.clone() let:row>
-            {contribution_row_view(&row)}
-        </For>
-    }
-    .into_any();
-    // `offset` is the offset that produced this page; the Suspend re-runs on any
-    // change (the list resource depends on it), so an untracked read is exact.
-    let current = offset.get_untracked();
-    let total = page.total;
-    let shown = u32::try_from(page.rows.len()).unwrap_or(u32::MAX);
-    let prev = (current > 0).then(|| {
-        view! {
-            <button
-                type="button"
-                class=BTN_SECONDARY
-                on:click=move |_| offset.set(current.saturating_sub(CONTRIBUTION_FETCH))
-            >
-                "← Previous"
-            </button>
-        }
-        .into_any()
-    });
-    let next = (current.saturating_add(shown) < total).then(|| {
-        view! {
-            <button
-                type="button"
-                class=BTN_SECONDARY
-                on:click=move |_| offset.set(current.saturating_add(CONTRIBUTION_FETCH))
-            >
-                "Next →"
-            </button>
-        }
-        .into_any()
-    });
-    view! {
-        {table_shell(&["Contribution", "Committed", "Committer", "Change type"], body)}
-        <p class="mt-2 text-xs text-ink-muted">{total} " contribution(s) total"</p>
-        <div class="mt-3 flex gap-2">{prev} {next}</div>
-    }
-    .into_any()
-}
-
-/// One contribution row: the uid (mono) plus its commit metadata.
-fn contribution_row_view(row: &ContributionRow) -> AnyView {
-    view! {
-        <tr class=ROW>
-            <td class=CELL_MONO>{row.uid.clone()}</td>
-            <td class=CELL>{row.time_committed.clone()}</td>
-            <td class=CELL>{row.committer.clone()}</td>
-            <td class=CELL>{row.change_type.clone()}</td>
-        </tr>
-    }
-    .into_any()
-}
-
-/// The by-uid CONTRIBUTION lookup box kept below the list: submitting a uid
-/// drives [`fetch_contribution`] under a `<Transition>` and renders the raw
-/// canonical JSON.
-fn contribution_lookup(ehr_id: Signal<String>, selected: Memo<String>) -> AnyView {
-    let uid_input = RwSignal::new(String::new());
-    let submitted = RwSignal::new(String::new());
-    let on_click = move |_| submitted.set(uid_input.get().trim().to_owned());
-    let resource = Resource::new(
-        move || {
-            let uid = submitted.get();
-            (selected.get() == "contributions" && !uid.is_empty()).then(|| (ehr_id.get(), uid))
-        },
-        |active| async move {
-            match active {
-                Some((id, uid)) => fetch_contribution(id, uid).await.map(Some),
-                None => Ok(None),
-            }
-        },
-    );
-    let lookup = view! {
-        <section class=format!("{CARD_PAD} mb-4")>
-            <div class="flex flex-wrap items-end gap-3">
-                <div class="flex flex-col gap-1">
-                    <label class=LABEL r#for="contribution-uid">
-                        "Contribution uid"
-                    </label>
-                    <input
-                        id="contribution-uid"
-                        type="text"
-                        class=INPUT
-                        placeholder="contribution uid (UUID)"
-                        prop:value=move || uid_input.get()
-                        on:input:target=move |ev| uid_input.set(ev.target().value())
-                    />
-                </div>
-                <button type="button" class=BTN_SECONDARY on:click=on_click>
-                    "Look up"
-                </button>
-            </div>
-        </section>
-    }
-    .into_any();
-    let result = view! {
-        <Transition fallback=table_skeleton>
-            {move || Suspend::new(async move {
-                match resource.await {
-                    Ok(Some(body)) => contribution_body(&body),
-                    Ok(None) => {
-                        // Resolve inside the Transition: an SSR'd ErrorBoundary fallback
-                        // mismatches at hydration in leptos 0.8 (E2E console gate).
-                        view! {
-                            <p class="text-sm text-ink-muted">
-                                "Enter a contribution uid to look it up."
-                            </p>
-                        }
-                            .into_any()
-                    }
-                    Err(e) => crate::components::format_view::inline_error(&e),
-                }
-            })}
-        </Transition>
-    }
-    .into_any();
-    view! { <div>{lookup} {result}</div> }.into_any()
-}
-
-/// Render a fetched CONTRIBUTION as pretty JSON in the shared document pane.
-fn contribution_body(body: &str) -> AnyView {
-    let pretty =
-        crate::components::format_view::pretty_body(body, crate::format::ReprFormat::CanonicalJson);
-    let doc_sig = RwSignal::new(pretty);
-    view! { <DocumentPane body=doc_sig /> }.into_any()
-}
-
-#[cfg(all(test, feature = "ssr"))]
-mod tests {
-    use super::{
-        DIRECTORY_ARCHETYPE, LIST_COMPOSITIONS_AQL, commit_version_uid, directory_toast_detail,
-        empty_root_folder, folder_json, format_from_value, format_value, parse_contributions,
-        versioned_object_id,
-    };
-    use crate::format::ReprFormat;
-
-    #[test]
-    fn folder_json_builds_a_spec_valid_folder_node() {
-        let leaf = folder_json(super::FOLDER_NODE_ID, "2026", &[]);
-        let root = folder_json(DIRECTORY_ARCHETYPE, "episodes", &[leaf]);
-        assert_eq!(root["_type"], "FOLDER");
-        assert_eq!(root["archetype_node_id"], DIRECTORY_ARCHETYPE);
-        assert_eq!(root["name"]["value"], "episodes");
-        assert_eq!(root["name"]["_type"], "DV_TEXT");
-        // Child folder present; no uid (the CDR assigns the OBJECT_VERSION_ID).
-        assert_eq!(root["folders"][0]["name"]["value"], "2026");
-        assert!(root.get("uid").is_none());
-        assert!(root["items"].as_array().unwrap().is_empty());
-    }
-
-    #[test]
-    fn empty_root_folder_is_an_empty_directory_root() {
-        let root = empty_root_folder();
-        assert_eq!(root["_type"], "FOLDER");
-        assert_eq!(root["archetype_node_id"], DIRECTORY_ARCHETYPE);
-        assert!(root["folders"].as_array().unwrap().is_empty());
-    }
-
-    #[test]
-    fn directory_toast_detail_names_the_version_or_a_generic_line() {
-        assert_eq!(
-            directory_toast_detail("7d44::sys::1"),
-            "New version 7d44::sys::1"
-        );
-        assert_eq!(directory_toast_detail(""), "The directory was committed.");
-    }
-
-    #[test]
-    fn parse_contributions_reads_rows_and_total_defensively() {
-        let body = r#"{
-            "rows": [
-                {
-                    "uid": "c1::sys::1",
-                    "time_committed": "2026-07-12T10:00:00Z",
-                    "committer": "Dr Bob",
-                    "change_type": "creation"
-                },
-                {"uid": "c2::sys::1"}
-            ],
-            "total": 42
-        }"#;
-        let page = parse_contributions(body).expect("valid contributions page");
-        assert_eq!(page.total, 42);
-        assert_eq!(page.rows.len(), 2);
-        assert_eq!(page.rows[0].uid, "c1::sys::1");
-        assert_eq!(page.rows[0].time_committed, "2026-07-12T10:00:00Z");
-        assert_eq!(page.rows[0].committer, "Dr Bob");
-        assert_eq!(page.rows[0].change_type, "creation");
-        // A row missing fields reads as empty strings, never a parse failure.
-        assert_eq!(page.rows[1].uid, "c2::sys::1");
-        assert_eq!(page.rows[1].time_committed, "");
-        assert_eq!(page.rows[1].committer, "");
-        assert_eq!(page.rows[1].change_type, "");
-    }
-
-    #[test]
-    fn parse_contributions_defaults_absent_rows_and_total() {
-        let page = parse_contributions("{}").expect("empty object parses");
-        assert!(page.rows.is_empty());
-        assert_eq!(page.total, 0);
-    }
-
-    #[test]
-    fn parse_contributions_rejects_non_json() {
-        assert!(parse_contributions("not json").is_err());
-    }
-
-    #[test]
-    fn fixed_aql_parses() {
-        openehr_query::parser::parse_str(LIST_COMPOSITIONS_AQL)
-            .expect("the compositions AQL const must parse");
-    }
-
-    #[test]
-    fn versioned_object_id_strips_the_version_suffix() {
-        assert_eq!(
-            versioned_object_id("7d44aa01::example.ehrbase.org::2"),
-            "7d44aa01"
-        );
-        // A bare versioned-object id (no suffix) is returned unchanged.
-        assert_eq!(versioned_object_id("7d44aa01"), "7d44aa01");
-    }
-
-    #[test]
-    fn commit_version_uid_reads_uid_value_or_empty() {
-        let body =
-            r#"{"_type":"COMPOSITION","uid":{"_type":"OBJECT_VERSION_ID","value":"7d44::sys::1"}}"#;
-        assert_eq!(commit_version_uid(body), "7d44::sys::1");
-        // A return=minimal (empty) or non-JSON body yields no uid.
-        assert_eq!(commit_version_uid(""), "");
-        assert_eq!(commit_version_uid("{}"), "");
-    }
-
-    #[test]
-    fn format_value_round_trips() {
-        for format in [
-            ReprFormat::CanonicalJson,
-            ReprFormat::CanonicalXml,
-            ReprFormat::Flat,
-        ] {
-            assert_eq!(format_from_value(format_value(format)), format);
-        }
-        // An unknown value falls back to canonical JSON.
-        assert_eq!(format_from_value("bogus"), ReprFormat::CanonicalJson);
-    }
-}
